back-end/requirements.txt
djangorestframework-gis==1.0
affine==2.4.0
amqp==5.1.1
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
billiard==4.1.0
bokeh==3.4.0.dev4
bqplot==0.12.42
branca==0.7.0
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
cffi==1.15.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
colour==0.1.5
comm==0.2.0
# contourpy==1.1.0
cssselect==1.1.0
decorator==5.1.1
Django==4.1.11
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
dnspython==2.4.2
drf-yasg==1.21.7
drigo==0.2.0
earthengine-api==0.1.367
eerepr==0.0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
folium==0.15.0
fonttools==4.42.1
geemap==0.29.5
geocoder==1.38.1
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
importlib-resources==6.0.1
inflection==0.5.1
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.17.4
ipython==8.17.2
ipytree==0.2.2
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jsonpointer==2.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
kombu==5.3.2
llvmlite
MarkupSafe==2.1.3
matplotlib
matplotlib-inline
munkres==1.1.4
numba
numpy
openpyxl==3.1.2
pandas
parso==0.8.3
pexpect==4.8.0
Pillow==10.0.0
plotly==5.18.0
ply==3.11
prompt-toolkit==3.0.39
protobuf==4.24.2
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.5.0
pyperclip==1.8.2
pyproj
PyQt5==5.15.10
pyquery==1.4.3
pyshp==2.3.1
python-box==7.1.1
pytz==2023.3.post1
PyYAML==6.0.1
rasterio==1.3.9
ratelim==0.1.6
rsa==4.9
scooby==0.9.2
shapely
snuggs==1.4.7
sqlparse==0.2.4
stack-data==0.6.3
tenacity==8.2.3
traitlets==5.13.0
traittypes==0.2.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
widgetsnbextension==4.0.9
xyzservices==2023.10.1
zipp==3.16.2
xarray
cdsapi
